metadata: Relocated some parsing-related functions from metadata.py to jinja_context.py (Did not change their contents.)
